Move `GlobalPath` approximation code to `path`
